Created toast for notification and news and  added notification toast in tab 1 and tab 2 of overview page.
